+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.1.2
 - Added an update button in settings to easily update the plugin when there's a new version available 🔃
 - Updated internal SDK to match latest version of RL (v2.55) 🚀
+- Added tooltips about enabling Quick Chats in order to view other players' titles ⚠️
 
 ## v1.1.1
 - Improved stability to reduce chance of crashing when joining a match